+++ test/libsolidity/ABITestsCommon.h
@@ -14,40 +14,30 @@
 	You should have received a copy of the GNU General Public License
 	along with solidity.  If not, see <http://www.gnu.org/licenses/>.
 */
-/**
- * @author Christian <c@ethdev.com>
- * @date 2016
- * Forward declaration of classes for inline assembly / JULIA AST
- */
 
-#pragma once
-
-#include <boost/variant.hpp>
+#include <string>
 
 namespace dev
 {
 namespace solidity
 {
-namespace assembly
+namespace test
 {
 
-struct Instruction;
-struct Literal;
-struct Label;
-struct StackAssignment;
-struct Identifier;
-struct Assignment;
-struct VariableDeclaration;
-struct FunctionalInstruction;
-struct FunctionDefinition;
-struct FunctionCall;
-struct If;
-struct Switch;
-struct ForLoop;
-struct Block;
+static std::string const NewEncoderPragma = "pragma experimental ABIEncoderV2;\n";
 
-using Statement = boost::variant<Instruction, Literal, Label, StackAssignment, Identifier, Assignment, FunctionCall, FunctionalInstruction, VariableDeclaration, FunctionDefinition, If, Switch, ForLoop, Block>;
+#define NEW_ENCODER(CODE) \
+{ \
+	sourceCode = NewEncoderPragma + sourceCode; \
+	{ CODE } \
+}
+
+#define BOTH_ENCODERS(CODE) \
+{ \
+	{ CODE } \
+	NEW_ENCODER(CODE) \
+}
 
 }
 }
-}
+} // end namespaces